# api/openapi/control-plane.yaml
openapi: 3.1.0
info:
  title: AJiaSu Control Plane API
  version: 1.0.0
  description: Phase 2 tenant, identity, and audit management API.
servers:
  - url: /
...
paths:
...
  /api/v1/tenants/{tenant_id}/endpoints:
    parameters: [{$ref: '#/components/parameters/TenantID'}]
    get:
      operationId: listEndpoints
      tags: [endpoints]
      security: [{cookieAuth: []}, {bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/PageSize'}]
      responses:
        '200': {description: Endpoint page., content: {application/json: {schema: {$ref: '#/components/schemas/EndpointList'}}}}
    post:
      operationId: createFixedEndpoint
      tags: [endpoints]
      security: [{cookieAuth: []}, {bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CSRFToken'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/EndpointCreate'}}}}
      responses:
        '202': {description: Endpoint desired state and queued operation., content: {application/json: {schema: {$ref: '#/components/schemas/EndpointOperation'}}}}
        '409': {$ref: '#/components/responses/Error'}
  /api/v1/tenants/{tenant_id}/endpoints/{endpoint_id}:
    parameters: [{$ref: '#/components/parameters/TenantID'}, {$ref: '#/components/parameters/EndpointID'}]
    get:
      operationId: getEndpoint
      tags: [endpoints]
      security: [{cookieAuth: []}, {bearerAuth: []}]
      responses:
        '200': {description: Endpoint desired and observed state., content: {application/json: {schema: {$ref: '#/components/schemas/Endpoint'}}}}
        '404': {$ref: '#/components/responses/Error'}
    patch:
      operationId: updateEndpointDesiredState
      tags: [endpoints]
      security: [{cookieAuth: []}, {bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CSRFToken'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [expected_version, desired_runner_state], properties: {expected_version: {type: integer, format: int64, minimum: 1}, desired_runner_state: {type: string, enum: [running, stopped]}}}}}}
      responses:
        '202': {description: Endpoint desired state queued., content: {application/json: {schema: {$ref: '#/components/schemas/EndpointOperation'}}}}
        '409': {$ref: '#/components/responses/Error'}
    delete:
      operationId: deleteEndpoint
      tags: [endpoints]
      security: [{cookieAuth: []}, {bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CSRFToken'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [expected_version], properties: {expected_version: {type: integer, format: int64, minimum: 1}}}}}}
      responses:
        '202': {description: Endpoint deletion queued., content: {application/json: {schema: {$ref: '#/components/schemas/EndpointOperation'}}}}
        '409': {$ref: '#/components/responses/Error'}
  /api/v1/tenants/{tenant_id}/endpoints/{endpoint_id}/start:
    parameters: [{$ref: '#/components/parameters/TenantID'}, {$ref: '#/components/parameters/EndpointID'}]
    post:
      operationId: startEndpoint
      tags: [endpoints]
      security: [{cookieAuth: []}, {bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CSRFToken'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/EndpointAction'}}}}
      responses: {'202': {description: Endpoint start queued., content: {application/json: {schema: {$ref: '#/components/schemas/EndpointOperation'}}}}, '409': {$ref: '#/components/responses/Error'}}
  /api/v1/tenants/{tenant_id}/endpoints/{endpoint_id}/stop:
    parameters: [{$ref: '#/components/parameters/TenantID'}, {$ref: '#/components/parameters/EndpointID'}]
    post:
      operationId: stopEndpoint
      tags: [endpoints]
      security: [{cookieAuth: []}, {bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CSRFToken'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/EndpointAction'}}}}
      responses: {'202': {description: Endpoint stop queued., content: {application/json: {schema: {$ref: '#/components/schemas/EndpointOperation'}}}}, '409': {$ref: '#/components/responses/Error'}}
  /api/v1/tenants/{tenant_id}/endpoints/{endpoint_id}/rebuild:
    parameters: [{$ref: '#/components/parameters/TenantID'}, {$ref: '#/components/parameters/EndpointID'}]
    post:
      operationId: rebuildEndpoint
      tags: [endpoints]
      security: [{cookieAuth: []}, {bearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CSRFToken'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/EndpointAction'}}}}
      responses: {'202': {description: Endpoint rebuild queued., content: {application/json: {schema: {$ref: '#/components/schemas/EndpointOperation'}}}}, '409': {$ref: '#/components/responses/Error'}}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: ajiasu_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 255}
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: false
      schema: {type: string}
      description: Required for cookie-authenticated unsafe requests.
    Cursor:
      name: cursor
      in: query
      schema: {$ref: '#/components/schemas/Cursor'}
    PageSize:
      name: page_size
      in: query
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
    TenantID:
      name: tenant_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
...
    EndpointID:
      name: endpoint_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
...
  responses:
...
    Error:
      description: Stable API error.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  schemas:
...
    EndpointStatus: {type: object, required: [observed_generation, observed_state, reason_code, last_transition_at], properties: {observed_generation: {type: integer, format: int64}, observed_state: {type: string, enum: [pending, starting, running, stopping, stopped, failed, orphaned]}, runner_id: {$ref: '#/components/schemas/UUID'}, reason_code: {type: string}, last_agent_observation_at: {$ref: '#/components/schemas/RFC3339Time'}, last_transition_at: {$ref: '#/components/schemas/RFC3339Time'}}}
    Endpoint: {type: object, required: [id, tenant_id, name, binding_mode, account_id, node_id, desired_runner_state, lifecycle_state, desired_generation, version, status, created_at, updated_at], properties: {id: {$ref: '#/components/schemas/UUID'}, tenant_id: {$ref: '#/components/schemas/UUID'}, name: {type: string}, binding_mode: {type: string, const: fixed}, account_id: {$ref: '#/components/schemas/UUID'}, node_id: {$ref: '#/components/schemas/UUID'}, desired_runner_state: {type: string, enum: [running, stopped]}, lifecycle_state: {type: string, enum: [active, disabled, deleting]}, desired_generation: {type: integer, format: int64}, version: {type: integer, format: int64}, status: {$ref: '#/components/schemas/EndpointStatus'}, created_at: {$ref: '#/components/schemas/RFC3339Time'}, updated_at: {$ref: '#/components/schemas/RFC3339Time'}}}
    EndpointCreate: {type: object, additionalProperties: false, required: [name, account_id, node_id], properties: {name: {type: string, minLength: 1, maxLength: 200}, account_id: {$ref: '#/components/schemas/UUID'}, node_id: {$ref: '#/components/schemas/UUID'}, desired_runner_state: {type: string, enum: [running, stopped], default: running}}}
    RunnerOperation: {type: object, required: [id, tenant_id, operation_type, resource_type, resource_id, requested_generation, state, progress_category, created_at, updated_at], properties: {id: {$ref: '#/components/schemas/UUID'}, tenant_id: {$ref: '#/components/schemas/UUID'}, operation_type: {type: string}, resource_type: {type: string}, resource_id: {$ref: '#/components/schemas/UUID'}, requested_generation: {type: integer, format: int64}, state: {type: string, enum: [queued, running, succeeded, failed, cancelled]}, progress_category: {type: string}, created_at: {$ref: '#/components/schemas/RFC3339Time'}, updated_at: {$ref: '#/components/schemas/RFC3339Time'}}}
    EndpointOperation: {type: object, required: [endpoint, operation], properties: {endpoint: {$ref: '#/components/schemas/Endpoint'}, operation: {$ref: '#/components/schemas/RunnerOperation'}}}
    EndpointAction: {type: object, additionalProperties: false, required: [expected_version], properties: {expected_version: {type: integer, format: int64, minimum: 1}}}
...
    EndpointList: {type: object, required: [items], properties: {items: {type: array, items: {$ref: '#/components/schemas/Endpoint'}}, next_cursor: {$ref: '#/components/schemas/Cursor'}}}
...
    UUID: {type: string, format: uuid}
    RFC3339Time: {type: string, format: date-time}
    Cursor: {type: string, minLength: 1, description: Opaque versioned base64url cursor.}
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id, details]
          properties:
            code: {type: string}
            message: {type: string}
            request_id: {$ref: '#/components/schemas/UUID'}
            details: {type: object, additionalProperties: true}
...